:sparkles: Exercicio: CRUD Cliente

+++ cap2/ecommerce/src/main/java/com/lucasengcomp/ecommerce/model/Cliente.java
@@ -9,22 +9,22 @@ import java.util.Objects;
 public class Cliente {
 
     @Id
-    private Long id;
+    private Integer id;
     private String nome;
 
     public Cliente() {
     }
 
-    public Cliente(Long id, String nome) {
+    public Cliente(Integer id, String nome) {
         this.id = id;
         this.nome = nome;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
